Extract Playwright and Cypress step details into separate files
